requirements.txt
numpy<1.23.0
einops
mujoco_py>=2.0
gym>=0.13,<0.24
dm_control>=1.0.3,<=1.0.20
cython==3.0.0a10
numba<0.60.0
matplotlib<=3.7.5
six
hydra-core
zarr<2.17
wandb
dill
av
pygame
pymunk
shapely<2.0.0
scikit-image<0.23.0
opencv-python
imagecodecs
mujoco<=3.1.6
